+++ src/base/__init__.py
@@ -1,0 +1,3 @@
+from .shipbase import ShipBase
+from .screenbase import ScreenBase
+from .menubase import MenuBase
