+++ src/sprites/__init__.py
@@ -1,4 +1,6 @@
-from .enemy import Enemy
-from .player import Player
 from .laser import Laser
+
+from .player import Player
+
+from .enemy import Enemy
 from .saucer import Saucer
